[not verified] Convert ProductOffer component to TypeScript
